Fine-tune the experiment setup process (#534)
* add get_required_env() to let setup_rocoto.py exit if a given env variable is NOT set
* each exp can have its own COMROOT and STMP under non-NCO environments
* fix the misuse of the reserved 'NET' variable
* each python file corresponds to one task
