+++ scripts/exrrfs_lbc.sh
@@ -35,7 +35,7 @@ elif  [[ "${prefix}" == "GEFS" ]]; then
 fi
 interval_seconds=3600 # just a place holder as we use metatask to run lbc hour by hour
 zeta_levels=${FIXrrfs}/meshes/L60.txt
-decomp_file_prefix="${NET}.graph.info.part."
+decomp_file_prefix="${MESH_NAME}.graph.info.part."
 #
 physics_suite=${PHYSICS_SUITE:-'PHYSICS_SUITE_not_defined'}
 file_content=$(< ${PARMrrfs}/${physics_suite}/namelist.init_atmosphere) # read in all content
@@ -48,8 +48,8 @@ sed -e "s/@input_stream@/init.nc/" -e "s/@output_stream@/foo.nc/" \
 #prepare for init_atmosphere
 ln -snf ${COMINrrfs}/${RUN}.${PDY}/${cyc}${ensindexstr}/ungrib_lbc/${prefix}:${start_time:0:13} .
 ln -snf ${COMINrrfs}/${RUN}.${PDY}/${cyc}${ensindexstr}/ic/init.nc .
-${cpreq} ${FIXrrfs}/meshes/${NET}.static.nc static.nc
-${cpreq} ${FIXrrfs}/graphinfo/${NET}.graph.info.part.${NTASKS} .
+${cpreq} ${FIXrrfs}/meshes/${MESH_NAME}.static.nc static.nc
+${cpreq} ${FIXrrfs}/graphinfo/${MESH_NAME}.graph.info.part.${NTASKS} .
 
 # run init_atmosphere_model
 source prep_step
