+++ scripts/exrrfs_ungrib.sh
@@ -59,7 +59,7 @@ CDATEout=$($NDATE ${FHR} ${CDATE})
 start_time=$(date -d "${CDATEout:0:8} ${CDATEout:8:2}" +%Y-%m-%d_%H:%M:%S) 
 end_time=${start_time}
 interval_seconds=3600
-if [[ "${NET}" == "conus3km"   ]]; then
+if [[ "${MESH_NAME}" == "conus3km"   ]]; then
   dx=3; dy=3
 else
   dx=12; dy=12
