+++ workflow/exp/exp.jet_conus12km
@@ -2,11 +2,12 @@
 # Aloha! This file is to setup some top level experiment options
 #
 export OPSROOT=/lfs5/BMC/nrtrr/NCO_dirs
-export EXP_BASEDIR=/home/role.rtrr/RRFS  # actual expdir name will be generated by setup_exp.py
-export NET=conus12km               # conus12km, conus12kmGSL, conus3km
-export EXP_NAME=${NET}             # if empty, expdir will be setup under ${version}/ directly
-export COMROOT=${OPSROOT}/com      # task input and output data as well as logs
-export DATAROOT=${OPSROOT}/stmp    # task workdirs ($DATA) which to be removed immediately upon task completion unless KEEPDATA=yes
+export NET=rrfs                    # rrfs, rtma, 
+export MESH_NAME=conus12km         # conus12km, conus3km, atl3km, etc
+export EXP_NAME=${MESH_NAME}
+export EXPDIR=/home/role.rtrr/RRFS/${VERSION}/${EXP_NAME}
+export COMROOT=${OPSROOT}/${EXP_NAME}/com      # task input and output data as well as logs
+export DATAROOT=${OPSROOT}/${EXP_NAME}/stmp    # task workdirs ($DATA) which to be removed immediately upon task completion unless KEEPDATA=yes
 export VERSION=1.0.1
 export TAG=c1v01
 #
