+++ workflow/exp/exp.setup
@@ -2,18 +2,20 @@
 # Aloha! This file is to setup some top level experiment options
 #
 export OPSROOT=/lfs5/BMC/wrfruc/gge/nco
-export EXP_BASEDIR=${OPSROOT}/exp  # actual expdir name will be generated by setup_exp.py
-export NET=conus12km               # conus12km, conus12kmGSL, conus3km
-export EXP_NAME=conus12km          # if empty, expdir will be setup under ${version}/ directly
-export COMROOT=${OPSROOT}/com      # task input and output data as well as logs
-export DATAROOT=${OPSROOT}/stmp    # task workdirs ($DATA) which to be removed immediately upon task completion unless KEEPDATA=yes
+export EXP_NAME=test1
+export NET=rrfs                    # rrfs, rtma, 
+export MESH_NAME=conus12km         # conus12km, conus3km, atl3km, etc
 export VERSION=1.0.1
 export TAG=c1v01
 #
+export EXPDIR=${OPSROOT}/exp/${VERSION}/${EXP_NAME}
+export COMROOT=${OPSROOT}/${EXP_NAME}/com      # task input and output data as well as logs
+export DATAROOT=${OPSROOT}/${EXP_NAME}/stmp    # task workdirs ($DATA) which to be removed immediately upon task completion unless KEEPDATA=yes
+#
 export REALTIME=false     # true, false
 export RETRO_PERIOD=2024052700-2024052800
-export RETRO_CYCLETHROTTLE=5
-export RETRO_TASKTHROTTLE=100
+export RETRO_CYCLETHROTTLE=2
+export RETRO_TASKTHROTTLE=50
 #
 export FCST_ONLY=${FCST_ONLY:-false}
 # preempt some settings
